requirements.txt
fastapi==0.115.14
sqlalchemy==2.0.41
uvicorn==0.35.0
pydantic==2.11.7
passlib[bcrypt]==1.7.4
bcrypt==4.1.2 
python-jose
python-multipart